requirements.txt
# Фреймворк для Telegram бота
python-telegram-bot


# Календарь для Telegram бота
python-telegram-bot-calendar


# Для работы с данными в табличном виде
pandas


# Движок для pandas, чтобы он мог читать и писать .xlsx файлы
openpyxl


# Для корректной работы с часовыми поясами
pytz


# Для чтения конфигурации (токенов, путей к БД) из .env файлов
python-dotenv


# Фреймворк для VK бота
vkbottle


# Для загрузки файлов в VK (отправка Excel-отчётов)
aiohttp